+++ TSynth/HWControls.h
@@ -77,11 +77,11 @@ ADC *adc = new ADC();
 #define DEBOUNCE 30
 
 static byte muxInput = 0;
-static int mux1ValuesPrev[MUXCHANNELS] = {};
-static int mux2ValuesPrev[MUXCHANNELS] = {};
+static uint16_t mux1ValuesPrev[MUXCHANNELS] = {};
+static uint16_t mux2ValuesPrev[MUXCHANNELS] = {};
 
-static int mux1Read = 0;
-static int mux2Read = 0;
+static uint16_t mux1Read = 0;
+static uint16_t mux2Read = 0;
 static int volumeRead = 0;
 static int volumePrevious = 0;
 static long encPrevious = 0;
